Support vm deploys

+++ script/deploy-arm.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
 
-# ./script/deploy-arm.sh mynetlinkid user1 "./test/8x16_128_byte 8 16"
+# ./script/deploy-arm.sh mynetlinkid user1 "./test/8x16_128_byte 8 16" -vm
 
 netlink_id=$1
 arm_user=$2 # There are 4 users on the machine: user1, user2, user3, user4
 execution_args=$3 # e.g. "./test/8x8_64_byte 8 8"
+mode=$4 # e.g. -arm or -vm
 
 root_dir=$( cd "$(dirname "${BASH_SOURCE[0]}")" || exit ; cd ../; pwd -P ) # https://stackoverflow.com/questions/24112727/relative-paths-based-on-file-location-instead-of-current-working-directory
 arm_password="q6coHjd7P"
@@ -33,39 +34,42 @@ commands+="make arm-neon;"
 commands+="chmod 777 ./main.exe;"
 commands+="echo;"
 
-# Transfer executable and test files to ARM machine
-commands+="echo 'Transfer files to ARM';"
-commands+="lftp -c \"open -u ${arm_user},${arm_password} ${arm_user}@arm;\
-  mkdir -p ${netlink_id}/test;\
-  put -O ${netlink_id} main.exe;\
-  mirror -R test ${netlink_id}/test\";"
-commands+="echo;"
+# Run on ARM machine
+if [[ "$mode" == "-arm" ]]; then
+  # Transfer executable and test files to ARM machine
+  commands+="echo 'Transfer files to ARM';"
+  commands+="lftp -c \"open -u ${arm_user},${arm_password} ${arm_user}@arm;\
+    mkdir -p ${netlink_id}/test;\
+    put -O ${netlink_id} main.exe;\
+    mirror -R test ${netlink_id}/test\";"
+  commands+="echo;"
 
-# Execute on ARM machine
-commands+="echo 'Execute on ARM';"
-commands+="eval \"{ sleep 1; echo ${arm_user};\
-  sleep 1; echo ${arm_password};\
-  echo 'cd ${netlink_id}';\
-  echo 'chmod 777 main.exe';\
-  echo './main.exe ${execution_args}';\
-  sleep 1; }\"\
-  | telnet arm;"
-commands+="echo;"
+  # Execute on ARM machine
+  commands+="echo 'Execute on ARM';"
+  commands+="eval \"{ sleep 1; echo ${arm_user};\
+    sleep 1; echo ${arm_password};\
+    echo 'cd ${netlink_id}';\
+    echo 'chmod 777 main.exe';\
+    echo './main.exe ${execution_args}';\
+    sleep 1; }\"\
+    | telnet arm;"
+  commands+="echo;"
 
-# Clean
-commands+="echo 'Clean up after deploy and execution';"
-commands+="cd ..;"
-commands+="rm -rf ${netlink_id};"
+  # Clean
+  commands+="echo 'Clean up after deploy and execution';"
+  commands+="cd ..;"
+  commands+="rm -rf ${netlink_id};"
 
-commands+="eval \"{ sleep 1;\
-  echo ${arm_user};\
-  sleep 1;\
-  echo ${arm_password};\
-  sleep 1;\
-  echo 'rmdir ${netlink_id}';\
-  sleep 1; }\"\
-  | telnet arm;"
-commands+="echo;"
+  commands+="eval \"{ sleep 1;\
+    echo ${arm_user};\
+    sleep 1;\
+    echo ${arm_password};\
+    sleep 1;\
+    echo 'rmdir ${netlink_id}';\
+    sleep 1; }\"\
+    | telnet arm;"
+  commands+="echo;"
+fi
 
 # commands+="bash;"
 
@@ -73,3 +77,14 @@ commands+="echo;"
 echo "ssh into remote @ seng440.ece.uvic.ca and execute commands";
 ssh -t "$netlink_id"@seng440.ece.uvic.ca "$commands";
 echo;
+
+# Run on local vm
+if [[ "$mode" == "-vm" ]]; then
+  echo "Download main.exe from remote @ seng440.ece.uvic.ca to local";
+  scp "$netlink_id"@seng440.ece.uvic.ca:/tmp/"$netlink_id"/main.exe "$root_dir";
+  echo;
+
+  echo "Send main.exe and test files to vm";
+  scp -P 5555 -r ./main.exe ./test root@localhost:~;
+  echo;
+fi
